ENH: revisited directory structure, reorganized and removed not anymore needed files
