--- CMake/vtkReportingVersionConfigure.h.in
@@ -1,19 +1,0 @@
-/*=auto=========================================================================
-
- Portions (c) Copyright 2005 Brigham and Women's Hospital (BWH)
- All Rights Reserved.
-
- See COPYRIGHT.txt
- or http://www.slicer.org/copyright/copyright.txt for details.
-
- Program:   3D Reporting
-
-=========================================================================auto=*/
-
-#ifndef __vtkReportingVersionConfigure_h
-#define __vtkReportingVersionConfigure_h
-
-#define Reporting_WC_REVISION "@Reporting_WC_REVISION@"
-#define Reporting_WC_URL "@Reporting_WC_URL@"
-
-#endif // __vtkReportingVersionConfigure_h
