+++ Plugins/DICOMTID1500Plugin.py
