+++ Plugins/base/DICOMPluginBase.py
