+++ QuantitativeReporting.py
