--- README
@@ -1,5 +1,0 @@
-This is the work in progress for the Slicer 4 Reporting module.
-
-Project details: http://wiki.na-mic.org/Wiki/index.php/Projects:QIN:3D_Slicer_Annotation_Image_Markup
-
-Documentation for the module: http://wiki.slicer.org/slicerWiki/index.php/Documentation/Nightly/Extensions/Reporting
