+++ SEGExporterSelfTest.py
