+++ startSlicelet.sh
